Move Public- and SecretKey to own module
